requirements.txt
fastapi
python-dotenv
uvicorn[standard]
google-cloud-speech
google-genai[asyncio]
python-multipart